requirements.txt
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0
neo4j>=5.14.0
graphiti-core>=0.3.0
streamlit>=1.28.0
pandas>=2.0.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
umap-learn>=0.5.0
python-dotenv>=1.0.0
